Finished Account Creation
Styling isn't done, but it functions like I want it to. Passwords can be visible or hidden, all fields are validated and cannot be empty, SSN auto-formats for validation and a success message is displayed when returning to the login page. There are also buttons to reset the fields and to return to the login page/cancel the creation!